WIP variable data type safety

+++ AS1 - Bouncing Balls - Visual Studio Project/AS1 - Bouncing Balls/GameGraphics.h
@@ -8,7 +8,7 @@
 
 class Cloud {
     float x_position;
-    float y_position;
+    int y_position;
     float x_scale;
     float y_scale;
     float x_velocity;
@@ -46,7 +46,7 @@ public:
 
     string raw_game_rules = "AIM AT BUBBLES OF THE SAME COLOR AND SHOOT!";
     string game_rules = "";
-    float bullet_radius = 8.0;
+    int bullet_radius = 8;
     int bullet_size = bullet_radius * 2;
     int bullet_mask_size = bullet_size * 1.25;
     int game_start_area = (Registry::window_size[0] - Registry::window_size[1]) / 2;
@@ -67,7 +67,7 @@ public:
 
     string raw_game_rules = "AIM AT BUBBLES OF THE SAME COLOR AND SHOOT!";
     string game_rules = "";
-    float bullet_radius = 8.0;
+    int bullet_radius = 8;
     int bullet_size = bullet_radius * 2;
     int bullet_mask_size = bullet_size * 1.25;
